Made a big friendly button for doing cloning...
The clone_picture is my own artwork, it reflects the obvious
lack of talent I possess as a graphic designer!

The 2 button nine-patch PNGs (btn_*_matte.9.png) are taken from this
blogpost by Folkert Jongbloed :

http://www.dibbus.com/2011/03/9patch-images-in-android/

He kindly says:

"Use it for your own projects. No, I don’t use any licence for this,
just use it and make tons of money with it!"