Fix #47 - Exception on checking out some repos due to InflaterInputStream
Calling InflaterInputStream.close() on Oracle Java doesn't call end() on
its Inflater if you supplied the Inflater in the constructor - but Android,
incorrectly, *does*. Inflaters can't be used after end() has been called,
which means you can't re-use Inflaters after using them with an
InflaterInputStream on Android - which is precisely what JGit tries to do
with InflaterCache. This results in either NullPointerException or
IllegalStateException depending on what version of Android you're using.

The fix here is to only use special HarmonyFixInflaterInflaters in
InflaterCache - these Inflaters can be re-used after being handed to
InflaterInputStream, because they ignore their end() method. This is ok
because you still have to hand your Inflater back to the InflaterCache
once you're finished with it, and InflaterCache will now 'decommision'
it's Inflaters - which thanks to our HARMONY_FIX_FACTORY actually
DOES end() the Inflater and release it's resources.

https://github.com/rtyley/agit/issues/47