upgrade to android-maven-plugin 3.2.0
This has been a long time coming - the proguard-handling changed with the
release of 3.0.0, and we've been stuck on an alpha-release ever since...
finally made it on to the modern version!

The exclusions are unfortunately necessary to avoid Dalvik verifier errors
as it freaks out over differently proguard-ed versions of the same class.