improve status-bar notifications for ICS and also Honeycomb
Legibility greatly improved, appropriate colours are used to match ordinary
notifications:

* ICS - the new Notification.Builder and the default notification layout
are used. The layout is abused slightly, with the 'info' textview being
updated (tho the id for this textview is hidden) to serve as the progress
text field.

* Honeycomb and below - differences in colour between different versions
of Android are handled by using the StatusBarNotificationStyles class,
which is based off http://stackoverflow.com/a/7320604/438886

The AndroidManifest target sdk update to '13' (ie Honeycomb) is necessary
to get the nice Holo-styling on the progress bar. Targeting ICS will have
to wait until ABS support is in:

"Applications targeting [14] or a later release will get these new changes
in behavior: For devices without a dedicated menu key, the software
compatibility menu key will not be shown even on phones. By targeting
Ice Cream Sandwich or later, your UI must always have its own menu UI
affordance if needed, on both tablets and phones. The ActionBar will take
care of this for you."