UX tweak, don't show fiddly fields until the Clone url is populated
Establishing the clone-url is always going to be the user's first priority,
so don't distract them with other fields until they've made headway on that
task.

This tweak is only enabled on Honeycomb and above, ie where the gentle
LayoutTransition animation is available. The change is too distracting if
it occurs instantly without animated fade-in.