Fix TagViewer regression and switch to integration test rather than robolectric
The awful guice repository-scope problems are avoided by passing the repo down
as a parameter - so the TagViewer is no longer blank.

The robolectric test has been dying horribly on some systems, seemly due to
bytecode problems, and it didn't give any indication of failure when the
TagViewer regressed - so an integration test is replacing the robolectric one.