Mildly improve error messages on OutOfMemoryError exceptions in Git ops
OutOfMemoryError is not an exception, so onException() was not getting
called. Also, the OutOfMemoryErrors I'm seeing generally don't have a
message, so change to use the exception name as a message until I can
think of something helpful to write about them...

Note that there is still more that can be done here - for one thing, the
sync adapter will keep on trying to hit the faulty repo, and will
consequently cause out of memory errors at will... need to store a flag
somewhere saying 'last fetch NOT successful' - and leave that remote alone.