Request a large Dalvik heap to help with cloning problematic repos
Thanks to @pmoor for pointing out the availability of this manifest
attribute, reporting issue #99!

Large objects are streamed wherever possible and consequently files of
this size often won't cause problems, even on a relatively memory
constrained device like an Android. That works so long as the large
object is stored as a whole object within the Git packfile. However, if
the file is similar to any other file stored anywhere within the
repositories history, they can both be stored in delta format - which
unfortunately requires a full in-memory representation to decompress.

https://github.com/rtyley/agit/wiki/Memory